+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Version 0.2.0 - Unreleased
 --------------------------
 
 * The controller form save is now working with basic fields;
+* Added working basic controller form;
+* Added model admins;
 
 
 Version 0.1.0 - Unreleased